requirements.txt
streamlit
langchain
openai
faiss-cpu
tiktoken
pypdf
duckduckgo-search
python-dotenv